#frontend
Filter functions for the sessions-dashboard have been extended:

One can now add a single subject to a session with a maximum of 25 symbols when creating or editing a session.

Subject-Filtering is now available on the session dashboard (multi-select dropdown).

If the session list is filtered, the amount of hidden sessions is now displayed at the end of the list. A click on that card resets the text search and subject filter to default, showing the original order again.

If a session has a subject attached, it will be shown in the session overview on the corresponding card.